Renamed ClusterProviderIconMenu file
