+++ shell/components/ClusterIconMenu.vue
